+++ limberframework/cache/stores.py
@@ -10,7 +10,6 @@ Functions:
 from abc import ABCMeta, abstractmethod
 from datetime import datetime
 from math import ceil
-from os.path import abspath
 from typing import Dict
 
 from aioredis import RedisConnection, create_redis
@@ -25,7 +24,7 @@ class Store(metaclass=ABCMeta):
     """Base class for a store."""
 
     @abstractmethod
-    def get(self, key: str) -> Dict:
+    async def get(self, key: str) -> Dict:
         """Retrieve data from cache.
 
         Arguments:
@@ -35,7 +34,7 @@ class Store(metaclass=ABCMeta):
         """
 
     @abstractmethod
-    def add(self, key: str, value: str, expires_at: datetime) -> bool:
+    async def add(self, key: str, value: str, expires_at: datetime) -> bool:
         """Store data in cache if it does not already exist.
 
         Arguments:
@@ -47,7 +46,7 @@ class Store(metaclass=ABCMeta):
         """
 
     @abstractmethod
-    def put(self, key: str, value: str, expires_at: datetime) -> bool:
+    async def put(self, key: str, value: str, expires_at: datetime) -> bool:
         """Store data in cache, overriding any existing data.
 
         Arguments:
@@ -138,7 +137,7 @@ class FileStore(Store):
         hasher = Hasher("sha1")
         return self.directory + "/" + hasher(key)
 
-    def get(self, key: str) -> Dict:
+    async def get(self, key: str) -> Dict:
         """Retrieves stored data for a key.
 
         Arguments:
@@ -163,7 +162,7 @@ class FileStore(Store):
             decoded_contents["value"], decoded_contents["expires_at"]
         )
 
-    def add(self, key: str, value: str, expires_at: datetime) -> bool:
+    async def add(self, key: str, value: str, expires_at: datetime) -> bool:
         """Add data to cache storage if it does not already exist.
 
         Arguments:
@@ -178,9 +177,9 @@ class FileStore(Store):
         if not FileSystem.has_file(path):
             return False
 
-        return self.put(key, value, expires_at)
+        return await self.put(key, value, expires_at)
 
-    def put(self, key: str, value: str, expires_at: datetime) -> bool:
+    async def put(self, key: str, value: str, expires_at: datetime) -> bool:
         """Add data to cache storage, overriding any existing data.
 
         Arguments:
@@ -203,14 +202,14 @@ class RedisStore(Store):
     def __init__(self, redis: Redis) -> None:
         self.redis = redis
 
-    def get(self, key: str) -> Dict:
+    async def get(self, key: str) -> Dict:
         contents = self.redis.get(key)
         return self.process(contents)
 
-    def add(self, key: str, value: str, expires_at: datetime) -> bool:
-        return self.put(key, value, expires_at, nx=True)
+    async def add(self, key: str, value: str, expires_at: datetime) -> bool:
+        return await self.put(key, value, expires_at, nx=True)
 
-    def put(
+    async def put(
         self, key: str, value: str, expires_at: datetime, **kwargs
     ) -> bool:
         contents = self.encode(value, expires_at)
@@ -251,19 +250,19 @@ class MemcacheStore(Store):
     def __init__(self, client: Client) -> None:
         self.client = client
 
-    def get(self, key: str) -> Dict:
+    async def get(self, key: str) -> Dict:
         contents = self.client.get(key)
         return self.process(contents)
 
-    def add(self, key: str, value: str, expires_at: datetime) -> bool:
+    async def add(self, key: str, value: str, expires_at: datetime) -> bool:
         contents = self.client.get(key)
 
         if contents:
             return False
 
-        return self.put(key, value, expires_at)
+        return await self.put(key, value, expires_at)
 
-    def put(self, key: str, value: str, expires_at: datetime) -> bool:
+    async def put(self, key: str, value: str, expires_at: datetime) -> bool:
         contents = self.encode(value, expires_at)
         number_seconds = ceil((expires_at - datetime.now()).total_seconds())
 
@@ -280,8 +279,7 @@ async def make_store(config: Dict) -> Store:
     Store object.
     """
     if config["driver"] == "file":
-        cache_path = abspath(config["path"])
-        return FileStore(cache_path)
+        return FileStore(config["path"])
     if config["driver"] == "redis":
         return RedisStore(
             Redis(