+++ tests/cache/test_store.py
@@ -64,6 +64,19 @@ async def test_make_store_invalid_driver():
     )
 
 
+@mark.asyncio
+async def test_make_store_asyncredis_without_password():
+    config = {
+        "driver": "asyncredis",
+        "host": "localhost",
+        "port": 1234,
+        "db": 0,
+    }
+
+    with raises(KeyError, match="password"):
+        await make_store(config)
+
+
 @mark.parametrize(
     "directory,path",
     [
@@ -85,89 +98,96 @@ def test_file_store_path(directory, path):
 
 
 @patch("limberframework.cache.stores.FileSystem")
-def test_file_store_get(mock_file_system):
+@mark.asyncio
+async def test_file_store_get(mock_file_system):
     date = datetime.now() + timedelta(seconds=60)
     value = "test"
     content = date.isoformat() + "," + value
     mock_file_system.read_file.return_value = content
 
     file_store = FileStore("/test")
-    response = file_store.get("test")
+    response = await file_store.get("test")
 
     assert response == {"data": value, "expires_at": date}
 
 
 @patch("limberframework.cache.stores.FileSystem")
-def test_file_store_get_expired(mock_file_system):
+@mark.asyncio
+async def test_file_store_get_expired(mock_file_system):
     date = datetime.now() - timedelta(seconds=60)
     value = "test"
     content = date.isoformat() + "," + value
     mock_file_system.read_file.return_value = content
 
     file_store = FileStore("/test")
-    response = file_store.get("test")
+    response = await file_store.get("test")
 
     assert response == {"data": None, "expires_at": None}
 
 
 @patch("limberframework.cache.stores.FileSystem")
-def test_file_store_get_invalid_file(mock_file_system):
+@mark.asyncio
+async def test_file_store_get_invalid_file(mock_file_system):
     mock_file_system.read_file.side_effect = FileNotFoundError()
 
     file_store = FileStore("/test")
-    response = file_store.get("test")
+    response = await file_store.get("test")
 
     assert response == {"data": None, "expires_at": None}
 
 
 @patch("limberframework.cache.stores.FileSystem")
-def test_file_store_put(mock_file_system):
+@mark.asyncio
+async def test_file_store_put(mock_file_system):
     key = "test"
     value = "test"
     expires_at = datetime.now()
 
     file_store = FileStore("/test")
-    response = file_store.put(key, value, expires_at)
+    response = await file_store.put(key, value, expires_at)
 
     assert response
     mock_file_system.write_file.assert_called_once()
 
 
 @patch("limberframework.cache.stores.FileSystem")
-def test_file_store_add_invalid_file(mock_file_system):
+@mark.asyncio
+async def test_file_store_add_invalid_file(mock_file_system):
     mock_file_system.has_file.return_value = False
     key = "test"
     value = "test"
     expires_at = datetime.now()
 
     file_store = FileStore("/test")
-    response = file_store.add(key, value, expires_at)
+    response = await file_store.add(key, value, expires_at)
 
     assert response is False
 
 
 @patch("limberframework.cache.stores.FileSystem")
-def test_file_store_add(mock_file_system):
+@mark.asyncio
+async def test_file_store_add(mock_file_system):
     mock_file_system.has_file.return_value = True
     key = "test"
     value = "test"
     expires_at = datetime.now()
 
     file_store = FileStore("/test")
-    response = file_store.add(key, value, expires_at)
+    response = await file_store.add(key, value, expires_at)
 
     assert response
 
 
 @patch("limberframework.cache.stores.FileSystem")
-def test_file_store_get_item(mock_file_system):
+@mark.asyncio
+async def test_file_store_get_item(mock_file_system):
     date = datetime.now() - timedelta(seconds=60)
     value = "test"
     content = date.isoformat() + "," + value
     mock_file_system.read_file.return_value = content
 
     file_store = FileStore("/test")
-    response = file_store["test"]
+    response = await file_store["test"]
 
     assert response == {"data": None, "expires_at": None}
 
@@ -237,18 +257,20 @@ def test_payload(data, expires_at, payload):
     assert response == payload
 
 
-def test_redis_store_get():
+@mark.asyncio
+async def test_redis_store_get():
     mock_redis = Mock()
     mock_redis.get.return_value = "2020-08-12T00:00:00,test".encode()
 
     redis_store = RedisStore(mock_redis)
-    response = redis_store.get("test")
+    response = await redis_store.get("test")
 
     assert response == {"data": "test", "expires_at": datetime(2020, 8, 12)}
 
 
 @patch("limberframework.cache.stores.datetime")
-def test_redis_store_add(mock_datetime):
+@mark.asyncio
+async def test_redis_store_add(mock_datetime):
     key = "test"
     value = "test"
     expires_at = datetime(2020, 8, 12, 1)
@@ -259,7 +281,7 @@ def test_redis_store_add(mock_datetime):
     mock_datetime.now.return_value = now
 
     redis_store = RedisStore(mock_redis)
-    response = redis_store.add(key, value, expires_at)
+    response = await redis_store.add(key, value, expires_at)
 
     assert response
     mock_redis.set.assert_called_once_with(
@@ -268,7 +290,8 @@ def test_redis_store_add(mock_datetime):
 
 
 @patch("limberframework.cache.stores.datetime")
-def test_redis_store_put(mock_datetime):
+@mark.asyncio
+async def test_redis_store_put(mock_datetime):
     key = "test"
     value = "test"
     expires_at = datetime(2020, 8, 12, 1)
@@ -279,7 +302,7 @@ def test_redis_store_put(mock_datetime):
     mock_datetime.now.return_value = now
 
     redis_store = RedisStore(mock_redis)
-    response = redis_store.put(key, value, expires_at)
+    response = await redis_store.put(key, value, expires_at)
 
     assert response
     mock_redis.set.assert_called_once_with(
@@ -287,18 +310,20 @@ def test_redis_store_put(mock_datetime):
     )
 
 
-def test_memcache_store_get():
+@mark.asyncio
+async def test_memcache_store_get():
     mock_memcache = Mock()
     mock_memcache.get.return_value = "2020-08-12T00:00:00,test".encode()
 
     memcache_store = MemcacheStore(mock_memcache)
-    response = memcache_store.get("test")
+    response = await memcache_store.get("test")
 
     assert response == {"data": "test", "expires_at": datetime(2020, 8, 12)}
 
 
 @patch("limberframework.cache.stores.datetime")
-def test_memcache_store_add(mock_datetime):
+@mark.asyncio
+async def test_memcache_store_add(mock_datetime):
     key = "test"
     value = "test"
     expires_at = datetime(2020, 8, 12, 1)
@@ -310,7 +335,7 @@ def test_memcache_store_add(mock_datetime):
     mock_datetime.now.return_value = now
 
     memcache_store = MemcacheStore(mock_memcache)
-    response = memcache_store.add(key, value, expires_at)
+    response = await memcache_store.add(key, value, expires_at)
 
     assert response
     mock_memcache.set.assert_called_once_with(
@@ -318,16 +343,20 @@ def test_memcache_store_add(mock_datetime):
     )
 
 
-def test_memcache_store_add_existing():
+@mark.asyncio
+async def test_memcache_store_add_existing():
     mock_memcache = Mock()
     memcache_store = MemcacheStore(mock_memcache)
-    response = memcache_store.add("test", "test", datetime(2020, 8, 12, 1))
+    response = await memcache_store.add(
+        "test", "test", datetime(2020, 8, 12, 1)
+    )
 
     assert not response
 
 
 @patch("limberframework.cache.stores.datetime")
-def test_memcache_store_put(mock_datetime):
+@mark.asyncio
+async def test_memcache_store_put(mock_datetime):
     key = "test"
     value = "test"
     expires_at = datetime(2020, 8, 12, 1)
@@ -338,7 +367,7 @@ def test_memcache_store_put(mock_datetime):
     mock_datetime.now.return_value = now
 
     memcache_store = MemcacheStore(mock_memcache)
-    response = memcache_store.put(key, value, expires_at)
+    response = await memcache_store.put(key, value, expires_at)
 
     assert response
     mock_memcache.set.assert_called_once_with(